% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fluorescence_deviation_injection.R
\name{add_fluorescence_deviation_injection}
\alias{add_fluorescence_deviation_injection}
\title{Add fluorescence deviation due to injection}
\usage{
add_fluorescence_deviation_injection(plate_layout)
}
\arguments{
\item{plate_layout}{a data frame generated by \code{seed_plate}}
}
\description{
Add fluorescence deviation due to injection
}
