% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mixing_wait_period.R
\name{add_mixing_wait_period}
\alias{add_mixing_wait_period}
\title{Add extra ticks corresponding to the mixing & wait period}
\usage{
add_mixing_wait_period(dat.sim.init, num.injections = 4)
}
\arguments{
\item{dat.sim.init}{a data frame produced by \code{seed_rawdata_table}}
}
\description{
Add extra ticks corresponding to the mixing & wait period
}
