% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ocr_deviation_injection.R
\name{add_ocr_deviation_injection}
\alias{add_ocr_deviation_injection}
\title{Add effects of injections on OCR}
\usage{
add_ocr_deviation_injection(stress_test = "mito")
}
\arguments{
\item{stress_test}{the stress test name, example "mito"}
}
\description{
Add effects of injections on OCR
}
