\name{adjust.length}
\alias{adjust.length}
\title{
Zero Padding
}
\description{
Add zeros to the end of the data if necessary so that 
its length is a power of 2.  It returns the data with zeros
added if nessary and the length of the adjusted data.
}
\usage{
adjust.length(inputdata)
}
\arguments{
\item{inputdata}{
either a text file or an S object containing
data.
}}
\value{
Zero-padded 1D array.
}
\details{
}
\references{
See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\seealso{
}
\keyword{ts}
\keyword{ts}
