\name{B4}
\alias{B4}
\title{Transient Signal}
\description{Transient signal.
}
\usage{data(B4)
}
\format{A vector containing 1024 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(B4)
plot.ts(B4)

}


\keyword{datasets}
