\name{back1.180}
\alias{back1.180}
\title{Acoustic Returns}
\description{Acoustic returns from ...
}
\usage{data(back1.180)
}
\format{A vector containing 7936 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}

\examples{

data(back1.180)
plot.ts(back1.180)

}

\keyword{datasets}
