\name{back1.220}
\alias{back1.220}
\title{Acoustic Returns}
\description{Acoustic returns from an underwater metallic object.
}
\usage{data(back1.220)
}
\format{A vector containing 7936 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(back1.220)
plot.ts(back1.220)

}


\keyword{datasets}
