\name{backscatter.1.180}
\alias{backscatter.1.180}
\title{Pixel from Amber Camara}
\description{Pixel from amber camara.
}
\usage{data(backscatter.1.180)
}
\format{A vector containing observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(backscatter.1.180)
plot.ts(backscatter.1.180)

}

\keyword{datasets}
