\name{click.asc}
\alias{click.asc}
\title{Pixel from Amber Camara}
\description{Pixel from amber camara.
}
\usage{data(click.asc)
}
\format{A vector containing observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(click.asc)
plot.ts(click.asc)

}

\keyword{datasets}
