% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ocr.R
\name{compare_ocr}
\alias{compare_ocr}
\title{OCR comparison between groups}
\usage{
compare_ocr(df, robust = TRUE, stats.only = FALSE)
}
\arguments{
\item{df}{a \code{data.frame} with raw data and produced by \code{import_seahorse}}

\item{robust}{a logical variable which indicates non-parametric tests will be used if it is TRUE by default.
when robust = FALSE, t test or anova will be used depending on the number of groups}

\item{stats.only}{a \code{logical} value that indicates whether only the statistical outcomes are returned}
}
\description{
compare_ocr compares OCR between experimental groups
}
\details{
When there are two groups, the Wilcoxon signed-rank test will be used for the group comparison.
When there are three groups, the Kruskal–Wallis one-way analysis of variance will be sued for the group comparison.
}
