% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_glycoapr.R
\name{convert_glycoapr}
\alias{convert_glycoapr}
\title{Quantify glycoAPR}
\usage{
convert_glycoapr(x, init_ocr, rot_ocr, which_assay)
}
\arguments{
\item{x}{a numeric value or a vector of ECAR}

\item{init_ocr}{a numeric value of initial OCR}

\item{rot_ocr}{a numeric value of OCR AA/rot}

\item{which_assay}{a \code{character} which can be "XFp", "XFe24", "XF" or "XFe96"}
}
\description{
This function calculates glycolytic ATP production rate based on ECAR
}
