% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mitoapr.R
\name{convert_mitoapr}
\alias{convert_mitoapr}
\title{Quantify mitoAPR}
\usage{
convert_mitoapr(x)
}
\arguments{
\item{x}{the rate of oxygen consumption that is coupled to ATP production during OXPHOS}
}
\description{
This function calculates mitochondrial ATP production rate based on OCR
}
