\name{D0}
\alias{D0}
\title{Transient Signal}
\description{Transient signal.
}
\usage{data(D0)
}
\format{A vector containing 1024 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(D0)
plot.ts(D0)

}

\keyword{datasets}
