\name{epl}
\alias{epl}
\title{
Plot Dyadic Wavelet Transform Extrema
}
\description{
Plot dyadic wavelet transform extrema (output of \code{\link{ext}}).
}
\usage{
epl(dwext)
}
\arguments{
\item{dwext}{
dyadic wavelet transform (output of \code{\link{ext}}).
}}
\value{

}
\details{
}
\references{
See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\seealso{
\code{\link{mw}}, \code{\link{ext}}, \code{\link{wpl}}.
}
\keyword{ts}
