\name{gsampleone}
\alias{gsampleone}
\title{
Sampled Identity
}
\description{
Generate a sampled identity matrix.
}
\usage{
gsampleone(node, scale, np)
}
\arguments{
\item{node}{
location of the reconstruction gabor functions.
}
\item{scale}{
scale of the gabor functions.
}
\item{np}{
size of the reconstructed signal.
}}
\value{
diagonal of the ``sampled'' \eqn{Q_1} term (1D vector)
}
\details{
}
\references{
See discussions in the text of "Time-Frequency Analysis''.
}
\seealso{
\code{\link{kernel}}, \code{\link{gkernel}.
}
\keyword{ts}
