% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seahorse.R
\name{import_seahorse}
\alias{import_seahorse}
\title{Import Seahorse data}
\usage{
import_seahorse(x)
}
\arguments{
\item{x}{Excel spreadsheet which contains the Seahorse XF assay result}
}
\description{
Load Seahorse Excel files into R as a data frame
}
\details{
Import the Excel spreadsheet which contains the Seahorse XF assay result
}
