\name{mbtrim}
\alias{mbtrim}
\title{
Trim Dyadic Wavelet Transform Extrema
}
\description{
Trimming of dyadic wavelet transform local extrema, using
bootstrapping.
}
\usage{
mbtrim(extrema, scale=FALSE, prct=0.95)
}
\arguments{
  \item{extrema}{
    dyadic wavelet transform extrema (output of \code{\link{ext}}).
  }
  \item{scale}{
    when set, the wavelet transform at each scale will be plotted 
    with the same scale.
  }
  \item{prct}{}
}
\value{
Structure containing
\item{original}{
  original signal.
}
\item{extrema}{
  trimmed extrema representation.
}
\item{Sf}{
  coarse resolution of signal.
}
\item{maxresoln}{
  number of decomposition scales.
}
\item{np}{
  size of signal.
}}
\details{
The distribution of extrema of dyadic wavelet transform at each scale is
generated by bootstrap method, and the 95\% critical value is used for
thresholding the extrema of the signal.
}
\references{
See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\seealso{
\code{\link{mntrim}}, \code{\link{mrecons}}, \code{\link{ext}}.
}
\keyword{ts}
