% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate_probe_position.R
\name{navigate_probe_position}
\alias{navigate_probe_position}
\title{Define probe position for every tick}
\usage{
navigate_probe_position(dat.sim.init, mixing_wait_period)
}
\arguments{
\item{dat.sim.init}{a data frame that is produced by \code{seed_rawdata_table}}

\item{mixing_wait_period}{a data frame that is produced by \code{add_mixing_wait_period}}
}
\description{
Define probe position for every tick
}
