\name{pixel_8.8}
\alias{pixel_8.8}
\title{Pixel from Amber Camara}
\description{Pixel from amber camara.
}
\usage{data(pixel_8.8)
}
\format{A vector containing observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(pixel_8.8)
plot.ts(pixel_8.8)

}

\keyword{datasets}
