% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fluorescence_drift.R
\name{qc_fluorescence_drift}
\alias{qc_fluorescence_drift}
\title{Quality control of fluorescence drift}
\usage{
qc_fluorescence_drift(dat)
}
\arguments{
\item{dat.raw}{a data frame with raw data and produced by \code{import_seahorse}}
}
\description{
Estimation of drift constants in temperature control wells
}
