% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_o2.R
\name{qc_o2}
\alias{qc_o2}
\title{O2 quality control}
\usage{
qc_o2(df, o2.threshold = 160)
}
\arguments{
\item{df}{a \code{data.frame} with raw data and produced by \code{import_seahorse}}

\item{o2.threshold}{a numeric value that an O2 level cannot be higher than}
}
\description{
This function flags wells with problematic O2 levels
}
