% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_ph.R
\name{qc_ph}
\alias{qc_ph}
\title{pH quality control}
\usage{
qc_ph(df, pH.threshold = 7)
}
\arguments{
\item{df}{a \code{data.frame} with raw data and produced by \code{import_seahorse}}

\item{pH.threshold}{a numeric value a pH value cannot be lower than}
}
\description{
This function flags wells with problematic pH
}
