\name{Undocumented}
\alias{PcaRidgeSampling}
\alias{RidgeDist}
\alias{RidgeIrregSampling}
\alias{RunRec}
\alias{SampleGen}
\alias{Sausage}
\alias{SpecGen}
\alias{band}
\alias{cgtRadar}
\alias{cloudXYZ}
\alias{confident}
\alias{crcirgrec}
\alias{crcirrec}
\alias{ddw}
\alias{dw}
\alias{fftshift}
\alias{girregrec}
\alias{hescrc}
\alias{irregrec}
\alias{mbpval}
\alias{mcgt}
\alias{mnpval}
\alias{pcacrc}
\alias{pcafamily}
\alias{pcamaxima}
\alias{pcamorwave}
\alias{pcarec}
\alias{pcaregrec}
\alias{pcaridrec}
\alias{pcazeroskeleton}
\alias{robustrec}
\alias{showRadar}
\alias{simplepcarec}
\alias{vwtTh}
\title{Undocumented Functions in Rwave}
\description{
  Numerous functions were not documented in the original Swave help files.
}
%\usage{}
%\arguments{}
%\value{}
%\details{}
\references{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
%\seealso{}
\keyword{ts}
