% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwave.R
\docType{package}
\name{rwave}
\alias{rwave}
\alias{rwave-package}
\title{rwave: Seahorse Wave in R}
\description{
A R package for Seahorse data visualization and statistical analysis
}
\author{
\strong{Maintainer}: Xiang Zhang \email{sc.xiang.zhang@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
