% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_rawdata_table.R
\name{seed_rawdata_table}
\alias{seed_rawdata_table}
\title{Initialization of data table for recording raw data}
\usage{
seed_rawdata_table(
  num.injections,
  tick.num = 15,
  tick.interval = 14,
  mixing_wait_duration = 120
)
}
\arguments{
\item{num.injections}{the number of injections used in Seahorse XF analysis}

\item{tick.num}{the number of ticks within each measurement period}

\item{tick.interval}{the time interval between two ticks in a measurement period, seconds}

\item{mixing_wait_duration}{the time interval between two measurement periods, seconds}
}
\description{
Initialization of data table for recording raw data
}
