% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_seahorse.R
\name{shiny_seahorse}
\alias{shiny_seahorse}
\title{Shiny app for Seahorse}
\usage{
shiny_seahorse(seahorse_rate_data)
}
\arguments{
\item{seahorse_rate_data}{Seahorse rate data produced by \code{import_seahorse}}
}
\description{
shiny_seahorse lauches a shiny app for visualization and communication of Seahorse analysis
}
