\name{signal_W_tilda.7}
\alias{signal_W_tilda.7}
\title{
    File from historical Swave package.  
}
\usage{data(signal_W_tilda.1)}
\description{
    The package maintainer believes this file was read or written by a C function (signal_W_tilda) called from \link{mrecons}, and was a precomputed kernel. All C function disk reads and writes have been disabled, but the files are preserved for historical purposes.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\seealso{
    \code{ \link{mrecons} }.
}
\keyword{datasets}
