% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_drift_well.R
\name{sim_drift_well}
\alias{sim_drift_well}
\title{Simulate fluorescent drift in temperature control wells}
\usage{
sim_drift_well(num.injections = 4)
}
\arguments{
\item{num.injections}{the number of injections}
}
\description{
Fluorescene intensity can vary over time in the blank wells
}
