% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_o2_well.R
\name{sim_o2_well}
\alias{sim_o2_well}
\title{Simulate the true O2 emission of a single well}
\usage{
sim_o2_well(ocr, tick.num = 15)
}
\arguments{
\item{ocr}{a vector of true biological oxygen consumption rates, pmol O2 / min}

\item{tick.num}{the number of ticks within each measurement period}
}
\description{
sim_o2_well simulates the O2 emission fluoresence of a single well
}
