% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_bioenergetic_space.R
\name{sketch_bioenergetic_space}
\alias{sketch_bioenergetic_space}
\title{Visualization of the bioenergetic space}
\usage{
sketch_bioenergetic_space(df, average = TRUE, background = FALSE)
}
\arguments{
\item{df}{a \code{data.frame} produced by \code{summarize_apr}}

\item{average}{a logical value indcates whether only to show the average of technical replicates}

\item{background}{a logical value indicates whetehr background is included in the plot}
}
\description{
This function plots mitoAPR against glycoAPR
}
