% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_comparison_apr.R
\name{sketch_comparison_apr}
\alias{sketch_comparison_apr}
\title{Visualize APR comparisons between groups}
\usage{
sketch_comparison_apr(df, which_apr)
}
\arguments{
\item{df}{a \code{data.frame} resulted by \code{compare_apr}}

\item{which_apr}{a \code{character} or a vector of \code{character} indicates the name(s) of APR to be compared}
}
\description{
\code{sketch_comparison_apr} creates a dot plot showing APR difference between groups
}
