% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_comparison_ocr.R
\name{sketch_comparison_ocr}
\alias{sketch_comparison_ocr}
\title{Visualize OCR comparisons between groups}
\usage{
sketch_comparison_ocr(df, which_ocr)
}
\arguments{
\item{df}{a \code{data.frame} resulted by \code{compare_ocr}}

\item{which_apr}{a \code{character} or a vector of \code{character} indicates the name(s) of OCR to be compared}
}
\description{
\code{sketch_comparison_ocr} creates a dot plot showing OCR difference between groups
}
