% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_ecar.R
\name{sketch_ecar}
\alias{sketch_ecar}
\title{Line plots of ECAR}
\usage{
sketch_ecar(df, trend = TRUE)
}
\arguments{
\item{df}{a \code{data.frame} produced by \code{import_seahorse}}

\item{trend}{a logical value indicates whether the median is used to represent the trend}
}
\value{
This function returns line plots of ECAR
}
\description{
This function draws line plots of EACR profile in different conditions
}
