% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_o2.R
\name{sketch_o2}
\alias{sketch_o2}
\title{Visualize the raw O2 pressure data}
\usage{
sketch_o2(df)
}
\arguments{
\item{df}{a \code{data frame} containing raw data loaded by \code{import_seahorse}}
}
\description{
sketch_o2 visualizes the raw measurements of O2 pressure
}
