% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_ocr.R
\name{sketch_ocr}
\alias{sketch_ocr}
\title{Line plots of OCR}
\usage{
sketch_ocr(df, which_level = "group", trend = TRUE)
}
\arguments{
\item{df}{a \code{data.frame} produced by \code{import_seahorse}}

\item{which_level}{a character either "well" or "group"}

\item{trend}{a logical value indicates whether the median of replicates will be used}
}
\value{
This function returns line plots of OCR
}
\description{
This function draws line plots of OCR profile in different conditions
}
