% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch_plate_layout.R
\name{sketch_plate_layout}
\alias{sketch_plate_layout}
\title{Visualize the experiment setup in a plate}
\usage{
sketch_plate_layout(experiment_setup)
}
\arguments{
\item{experiment_setup}{a data.frame with two columns, group label and number of replicates}
}
\description{
sketch_plate_layout shows the group assignment in every well
}
