\name{smoothts}
\alias{smoothts}
\title{
Smoothing Time Series
}
\description{
Smooth a time series by averaging window.
}
\usage{
smoothts(ts, windowsize)
}
\arguments{
\item{ts}{
Time series.
}
\item{windowsize}{
Length of smoothing window.
}}
\value{
Smoothed time series (1D array).
}
\details{
}
\references{
See discussions in the text of ``Time-Frequency Analysis''.
}
\seealso{
}
\keyword{ts}
