% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_assay.R
\name{summarize_assay}
\alias{summarize_assay}
\title{Summary of assay settings}
\usage{
summarize_assay(dat, what)
}
\arguments{
\item{dat}{a list produced by \code{import_seahorse}}

\item{what}{a character "measurement" or "tick"}
}
\description{
summarize_assay returns assay settings such as number of ticks per measurements, number of measurements, etc.
}
