% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fluorescence_drift.R
\name{summarize_fluorescence_drift}
\alias{summarize_fluorescence_drift}
\title{Summarize fluoresence drift}
\usage{
summarize_fluorescence_drift(df, num.ticks, stats.only = TRUE)
}
\arguments{
\item{df}{a data frame produced by \code{qc_fluorescence_drift}}

\item{num.ticks}{the number of ticks in a measurement period produced by \code{summarize_assay}}

\item{stats.only}{a logical value whether only the median of temperature controlled wells are shown}
}
\description{
Summarize fluoresence drift
}
