% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_seahorse.R
\name{summarize_seahorse}
\alias{summarize_seahorse}
\title{Summarize Sehorse in few parameters}
\usage{
summarize_seahorse(df, which_assay)
}
\arguments{
\item{df}{a data frame produced by \code{import_seahorse}}

\item{which_assay}{a character which can be "XFp", "XFe24", "XF" or "XFe96"}
}
\description{
summarize_seahorse summarizes Seahorse analysis in few parameters
}
