% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_well.R
\name{tag_well}
\alias{tag_well}
\title{Assign experiment group to a well}
\usage{
tag_well(experiment_setup)
}
\arguments{
\item{experiment_setup}{a data.frame with two columns, group label and number of replicates}
}
\description{
Assign experiment group to a well
}
