\name{vecgabor}
\alias{vecgabor}
\title{
Gabor Functions on a Ridge
}
\description{
Generate Gabor functions at specified positions on a ridge.
}
\usage{
vecgabor(sigsize, nbnodes, location, frequency, scale)
}
\arguments{
\item{sigsize}{
Signal size.
}
\item{nbnodes}{
Number of wavelets to be generated.
}
\item{location}{
b coordinates of the ridge samples (1D array of length nbnodes).
}
\item{frequency}{
frequency coordinates of the ridge samples (1D array of length nbnodes).
}
\item{scale}{
size parameter for the Gabor functions.
}}
\value{
size parameter for the Gabor functions.
}
\details{
}
\references{
}
\seealso{
\code{\link{vecmorlet}}.
}
\keyword{ts}
