\name{W_tilda.2}
\alias{W_tilda.2}
\title{Pixel from Amber Camara}
\description{Pixel from amber camara.
}
\usage{data(W_tilda.2)
}
\format{A vector containing observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998, eBook ISBN:978008053942), eBook ISBN:978008053942)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\examples{

data(W_tilda.2)
plot.ts(W_tilda.2)

}


\keyword{datasets}
