% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fluoresence_deviation_injection.R
\name{add_fluoresence_deviation_injection}
\alias{add_fluoresence_deviation_injection}
\title{Add fluoresence deviation due to injection}
\usage{
add_fluoresence_deviation_injection(plate_layout, num.injections = 4)
}
\arguments{
\item{plate_layout}{a data frame generated by \code{seed_plate}}

\item{num.injections}{the number of injections in the Seahorse XF analysis}
}
\description{
Add fluoresence deviation due to injection
}
