% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_true_ocr.R
\name{add_true_ocr}
\alias{add_true_ocr}
\title{Add true OCR to every well in a plate}
\usage{
add_true_ocr(plate_layout, baseline_ocr, num.injections = 4)
}
\arguments{
\item{plate_layout}{is a data.frame produced by \code{seed_plate}}

\item{baseline_ocr}{is a vector of true biological OCR without any injection}
}
\description{
add_true_ocr spikes in true biological OCR in every well
}
