% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seahorse.R
\name{import_seahorse}
\alias{import_seahorse}
\title{Import Seahorse data}
\usage{
import_seahorse(x)
}
\arguments{
\item{x}{Excel spreadsheet which contains the Seahorse XF assay result}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item Measurement
\item Well
\item Group
\item Time
\item OCR
\item ECAR
\item PER
}
#' Import Seahorse data

Load Seahorse Excel files into R as a data frame

Import the Excel spreadsheet which contains the Seahorse XF assay result

This function returns a \code{data.frame} including columns:
\itemize{
\item Measurement
\item Well
\item Group
\item Time
\item OCR
\item ECAR
\item PER
}
#' Import Seahorse data

Load Seahorse Excel files into R as a data frame

Import the Excel spreadsheet which contains the Seahorse XF assay result

This function returns a \code{data.frame} including columns:
\itemize{
\item Measurement
\item Well
\item Group
\item Time
\item OCR
\item ECAR
\item PER
}
#' Import Seahorse data

Load Seahorse Excel files into R as a data frame

Import the Excel spreadsheet which contains the Seahorse XF assay result

This function returns a \code{data.frame} including columns:
\itemize{
\item Measurement
\item Well
\item Group
\item Time
\item OCR
\item ECAR
\item PER
}
#' Import Seahorse data

Load Seahorse Excel files into R as a data frame

Import the Excel spreadsheet which contains the Seahorse XF assay result

This function returns a \code{data.frame} including columns:
\itemize{
\item Measurement
\item Well
\item Group
\item Time
\item OCR
\item ECAR
\item PER
}

import_seahorse("seahorse_c2c12.xlsx")

This function returns a \code{data.frame} including columns:
\itemize{
\item Measurement
\item Well
\item Group
\item Time
\item OCR
\item ECAR
\item PER
}
}
\description{
Load Seahorse Excel files into R as a data frame
}
\details{
Import the Excel spreadsheet which contains the Seahorse XF assay result
}
