% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_plate.R
\name{seed_plate}
\alias{seed_plate}
\title{Initialize the Seahorse XF output}
\usage{
seed_plate(experiment_setup)
}
\arguments{
\item{experiment_setup}{a data frame with two columns Group and n}
}
\description{
seed_plate produces a data frame with two columns Well and group Tag.
}
