% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_apr.R
\name{summarize_apr}
\alias{summarize_apr}
\title{Quantify ATP production rate}
\usage{
summarize_apr(df, which_assay)
}
\arguments{
\item{df}{a \code{data.frame} produced by \code{import_seahorse}}

\item{which_assay}{a \code{character} which can be "XFp", "XFe24", "XF" or "XFe96"}
}
\description{
This function quantifies glycoAPR and mitoAPR based on OCR and ECAR
}
