% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ocr.R
\name{summarize_ocr}
\alias{summarize_ocr}
\title{Representive OCR values in mito stress test}
\usage{
summarize_ocr(df)
}
\arguments{
\item{df}{a \code{data.frame} produced by \code{import_seahorse}}
}
\description{
This function returns a set of OCR values describing mito stress test
}
